+++ config/config.go
@@ -5,8 +5,9 @@ import (
 	"github.com/giantswarm/micrologger"
 	kingpin "gopkg.in/alecthomas/kingpin.v2"
 
-	"github.com/giantswarm/admission-controller/pkg/awsmachinedeployment"
-	"github.com/giantswarm/admission-controller/pkg/g8scontrolplane"
+	"github.com/giantswarm/admission-controller/pkg/aws/awscontrolplane"
+	"github.com/giantswarm/admission-controller/pkg/aws/awsmachinedeployment"
+	"github.com/giantswarm/admission-controller/pkg/aws/g8scontrolplane"
 )
 
 const (
@@ -20,6 +21,7 @@ type Config struct {
 	AvailabilityZones string
 
 	G8sControlPlane      g8scontrolplane.Config
+	AWSControlPlane      awscontrolplane.Config
 	AWSMachineDeployment awsmachinedeployment.Config
 }
 
@@ -39,12 +41,18 @@ func Parse() (Config, error) {
 	kingpin.Flag("tls-cert-file", "File containing the certificate for HTTPS").Required().StringVar(&result.CertFile)
 	kingpin.Flag("tls-key-file", "File containing the private key for HTTPS").Required().StringVar(&result.KeyFile)
 	kingpin.Flag("address", "The address to listen on").Default(defaultAddress).StringVar(&result.Address)
-	kingpin.Flag("availability-zones", "List of AWS availability zones.").Required().StringVar(&result.G8sControlPlane.ValidAvailabilityZones)
+	kingpin.Flag("availability-zones", "List of AWS availability zones.").Required().StringVar(&result.AvailabilityZones)
 
 	// add logger to each admission handler
 	result.G8sControlPlane.Logger = newLogger
+	result.AWSControlPlane.Logger = newLogger
 	result.AWSMachineDeployment.Logger = newLogger
 
 	kingpin.Parse()
+
+	// add availability zones to admitter configs
+	result.AWSControlPlane.ValidAvailabilityZones = result.AvailabilityZones
+	result.G8sControlPlane.ValidAvailabilityZones = result.AvailabilityZones
+
 	return result, nil
 }